Enable dynamically adding tstore maps [HZ-4290] (#670)
Enable dynamically adding map configuration. There are two ways for
adding a dynamic map configuration:
1) with the broadcasting `DynamicConfigurationAwareConfig#addMapConfig`,
and
2) with reloading the configuration via
`ClusterWideConfigurationService#update`.

Prior to this change, 1) was disabled while 2) - used by our cloud
offering - was enabled. This change brings the two approaches to a
common ground and enables both. There is still one exception though:
`DynamicConfigurationAwareConfig#addDeviceConfig` still throws
`UnsupportedOperationException`, while
`ClusterWideConfigurationService#update` allows adding (second and more)
devices. This inconsistency can be ironed out in a separate change. In
practice, this behavior limits the usability of 1) to referring to the
devices present in the static configuration.

To enable dynamic TS maps, the startup condition enabling TS service had
to bechanged. Prior to this change, the service got initialized only if
a TS map was present in the configuration. This forced our cloud
offering to define a "fake" TS map in the static configuration. Starting
with this change, the TS service can be initialized with
`-Dhazelcast.tiered.store.force.enabled=true`.

GitOrigin-RevId: 65f7edfcc8ce70f210f5b1b5affe3971ceee62e1