Persist dynamic indexes on new cluster members [HZ-4428] (#829)
This is a temporary solution to the problem: each time dynamic index
(one created by `addIndex`) is received during partition migration, we
ensure that it is persisted on receiving member, just in case if that is
a new index. The main case when this can happen is when a new member
joins the cluster.

The better solution would be to send dynamic indexes in PreJoin
operation but that will be a more complicated change.

This PR also fixes flaky `MapConfigurationIndexHotRestartTest`.

Followup to #596
Fixes [HZ-4428](https://hazelcast.atlassian.net/browse/HZ-4428)
Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/7055

GitOrigin-RevId: f939261de48d060c50d1cf2ed088e6ad4839e036