Fix NPE in AggregateOperations.maxBy when one member did not have any data [HZ-4467] (#895)
Null values are not allowed as pipeline items, so comparator used in
`maxBy` is not supposed to handle them. However, it was possible that
`maxBy` combine step could receive empty `MutableReference` if some node
did not have any items to aggregate - this is the initial value of the
accumulator.

Fixes HZ-4467
[Community slack
report](https://hazelcastcommunity.slack.com/archives/C013LM6JK7S/p1708478988464329)

GitOrigin-RevId: 84b60d5c0854932db71aa43215baed0df8fb25c1