Fix race condition when cancelling Jet jobs during initialization [HZ-4449] (#889)
The problem is hard to reproduce reliably. It helps to introduce delays
in `JobExecutionService.terminateExecution0` (after `terminateExecution`
invocation) and during job initialization (after `ExecutionContext` is
added to `executionContexts` map but before `SnapshotContext` creation,
but not on all members). It also helps to cancel the job manually in
addition to automatic cancellation after the test.

Fixes https://github.com/hazelcast/hazelcast/issues/26255

GitOrigin-RevId: cb15717b8a8a3d4870ba0d455d3ea4b969d34474