OnJoinCacheOperationTest Mockito + Azul potential fix [HZ-4453] (#908)
Closes: https://hazelcast.atlassian.net/browse/HZ-4453

This is a possible solution. One test remains where a static method is
mocked.

1) The problematic test identified in the logs is
"test_cachePostJoinOperationSucceeds_whenJCacheAvailable_noWarningIsLogged".
In this scenario, the static method is expected to return true.
Tests where the static method should return true have been modified to
eliminate the use of Mockito.
2) The mock of the static method remains only in the test where it is
expected that the static method will return false. According to the
GitHub issue, there were no issues with this test.
3) Looks like the test -
test_cachePostJoinOperationFails_whenJCacheNotAvailable_withCacheConfigs
was wrong, so it was fixed.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: b70c7c5fd4b48230344ffba9f2b533b5743e075f