Mock interface instead of class in MapWanSyncRESTTest [HZ-1993] [5.3.z] (#939)
Backport of https://github.com/hazelcast/hazelcast-mono/pull/845

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/5662,
[HZ-1993](https://hazelcast.atlassian.net/browse/HZ-1993)

The problem is that Mockito occasionally fails to mock the class and
instead calls the real method implementation.
To address this, let's try to mock the interface instead of the class.
By mocking the interface, the issue of invoking real methods is avoided.

To be able to mock the interface the following was done:
was added the used getSyncResult method to the WanReplicationService
interface.
WanSyncStateResult now is an interface located in the os rep.
EnterpriseWanSyncStateResult is an enterprise implementation of
WanSyncStateResult located in ee rep.

So we can mock WanReplicationService and WanSyncStateResult interfaces
instead of EnterpriseWanReplicationService and
EnterpriseWanSyncStateResult.

[HZ-1993]:
https://hazelcast.atlassian.net/browse/HZ-1993?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: oshultseva <olga.shultseva@hazelcast.com>
GitOrigin-RevId: 37f9d6a05c3786779e2c4945acc13fb7853cabed