NLCModeLicenseTest Mockito + Azul fix [HZ-4483] (#943)
https://hazelcast.atlassian.net/browse/HZ-4483

Several tests are failing only on Azul Prime JDKs due to an
incompatibility between it and Mockito.
The solution chosen was to exclude, if possible, the use of mock static
methods, MockedConstruction, and mock complex class.
To address this, refactoring was performed:
* BuiltInLicenseHolder has been implemented to receive and subsequently
provide built-in licenses to requesters.
Rather than utilizing a mock license, the reflection is used to modify
the license value within the license holder.
* LicenseHelper method call checks have been replaced with similar ones

Key point about the system: The built-in license remains constant
throughout the runtime and cannot be altered dynamically.
There exist two built-in license options: an empty license and NLC.
The configuration of the built-in license occurs during the build phase,
determined by the inclusion or exclusion of the NoLicenseProvider in the
Maven profile.
If the NoLicenseProvider is included, the built-in license is set to
null; otherwise, it is set to NLC.
The verification of the built-in license only takes place during
startup. So even if it changes somehow during runtime, then this will
not take any effect.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: bafe78e186553db714e44bd068d95925faeca651