Fix NPE from SecurityContext startup order, add tests [5.4.0] (#1070)
Backport of: https://github.com/hazelcast/hazelcast-mono/pull/1062

It was possible to encounter an NPE when configuring a
`SecurityInterceptorConfig` statically with a `className` definition -
this is because the `SecurityContextImpl` would attempt to fetch the
`UserCodeNamespaceService` for obtaining the default `ClassLoader`, but
it was attempting to retrieve a reference that wasn't available yet.

This commit resolves this by fetching the `UserCodeNamespaceService` via
the `NodeExtension` which has the service available. This commit also
adds regression tests for this scenario which was not covered before.

Fixes https://hazelcast.atlassian.net/browse/HZ-4578

GitOrigin-RevId: f133a9420602af0f68013cab2014550737f6f4a8