Validate MapConfig success in UpdateMapConfigOperation [HZ-4602] (#1130)
Currently, if the `UpdateMapConfigOperation` is invoked with an invalid
`WanReplicationRef` (containing a merge policy class name which is
invalid), then the `MapConfig` is updated for the `MapContainer` owning
the `IMap`, but the reinitialization of the WAN context will have
failed, so the `MapConfig` is now inconsistent.

This commit resolves this by making sure no exceptions are thrown while
applying the config update, and if they are rolling back to the original
MapConfig.

Fixes https://hazelcast.atlassian.net/browse/HZ-4602

GitOrigin-RevId: 339a45db25a7fbad50831efef9bd4b79a26546f8