Potential Mockito/Azul issue fix for `CompactStreamSerializerTest` [HZ-4568] (#1087)
[Several tests are
failing](https://hazelcast.atlassian.net/browse/HZ-2245) only on Azul
Prime JDKs due to an incompatibility between it and Mockito.

In https://github.com/hazelcast/hazelcast-mono/pull/845 it was
discovered that a causal factor was mocking classes, not interfaces.

To address this, refactored the test to use a real instantiation of
`CompactStreamSerializer` instead of a mock.
Also refactored to use a lazily-produced iterable as callers typically
only ever access the first element.

*Unable* to reproduce this issue locally to test if this will resolve
the issue.

Fixes: [HZ-4568](https://hazelcast.atlassian.net/browse/HZ-4568),
https://github.com/hazelcast/hazelcast/issues/26280

[HZ-4568]:
https://hazelcast.atlassian.net/browse/HZ-4568?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: fa50b9ec5c0a407c0fdf722e913a16724597fa78