Potential Mockito/Azul issue fix for `CompactSchemaReplicationSlowTest` [HZ-4566][API-2218] (#1133)
[Several tests are
failing](https://hazelcast.atlassian.net/browse/HZ-2245) only on Azul
Prime JDKs due to an incompatibility between it and Mockito.

Although in most cases we can replace a mocked concrete class with a
mocked interface instead, here it's the `spy` functionality being used
on real instances and there's no such workaround.

To address this:
- Refactored `CompactSchemaReplicationSlowTest` and
`CompactSchemaReplicationTest` to move replicated logic into parent
`CompactSchemaReplicationTestBase`
- _Where possible_ replaced checking whether a method was called (via
`spy`) to instead check the effect of that method
- E.G. rather than checking `onSchemaAckRequest` is never called,
instead check that the `replicationStatus` != `REPLICATED`, which is the
effect of the method being called.
- this cannot be done in every scenario (e.g. asserting that a method is
called exactly 100 times when it has no side effect), and while it
_would_ be possible to replicate the `spy` functionality with some
invocation-counting delegate wrapper, as the tests pass this doesn't
_appear_ to be necessary
- Finally, understanding that checking a method is/isn't called is
equivalent to checking the `replicationStatus`, identified some of the
assertions were duplicated and redundant - now removed

I *have* been able to [test this fix in a Jenkins
job](https://hazelcast.atlassian.net/browse/HZ-4566?focusedCommentId=95093)
to give some confidence it addresses the issue.

Fixes: [HZ-4566](https://hazelcast.atlassian.net/browse/HZ-4566),
[API-2218](https://hazelcast.atlassian.net/browse/API-2218),
https://github.com/hazelcast/hazelcast/issues/26245

[HZ-4566]:
https://hazelcast.atlassian.net/browse/HZ-4566?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 79092b41a4bc0c870f47a6fddd6cd73c41dc7682