Replace `NodeEngineImpl` mocks with mocks of parent `NodeEngine` interface where possible (#1156)
[Several tests are
failing](https://hazelcast.atlassian.net/browse/HZ-2245) only on Azul
Prime JDKs due to an incompatibility between it and Mockito.

In https://github.com/hazelcast/hazelcast-mono/pull/845 it was
discovered that a causal factor was mocking classes, not interfaces.

As such, where possible, `NodeEngineImpl` mocks should be replaced with
mocks of parent `NodeEngine` interface.

This does not target a particular test failure and is a cleanup to
prevent occuring in future.

GitOrigin-RevId: 44be1563736dbffba8400b356f18181c8012f546