Don't use the default password from the SymmetricEncryptionConfig [HZ-4402] (#1311)
This commit removes usage of the default password in the
`SymmetricEncryptionConfig`.
Encryption as a security feature doesn't make sense when used with a
default key/password.
The feature has been deprecated since Hazelcast IMDG `4.2`.

The constant with the original value of the default password is kept in
place as it's part of the public API. It's not used anymore for the
configuration. A weak password checker still uses it.

Breaking changes (list specific methods/types/messages):
* removes the default value for the `password` option in the
`SymmetricEncryptionConfig` class

---------

Co-authored-by: Jack Green <JackPGreen@Gmail.com>
GitOrigin-RevId: 8a94397d28a5afeb60b77bf16dba44b91c7214af