Implement VectorCollection.putAllAsync [HZAI-25] (#1302)
Adds VectorCollection.putAllAsync very similar to IMap implementation.
There are small differences with IMap:

1. member-side proxy supports batching configuration (currently via
method in `VectorCollectionProxy`)
2. batching is supported in async invocation and is able to send
requests to many members in parallel (unlike IMap), but does not have
any way of limiting of in-flight batches, so should be considered as
experimental and potentially causing OOMs (but not much more than an
attempt to send the same big collection in one shot without batching)
3. client-side proxy does not support batching and sends one message per
partition (not per member - like IMap) but has a small optimisation of
sending `set` instead of `putAll` if there is only a single entry. This
however may affect some statistics of operations.
4. serialization of `VectorEntries` is rolling-upgrade-friendly

In addition, `putAll` for for vector collection supports offloading
long-running operations (configurable max continuous execution time).
The idea is similar to IMap StepEngine, but extremely simplified and not
generic. Modifications of the collection are still executed on partition
thread, but after the threshold time elapses, subsequent execution is
scheduled on the same partition thread to free it in order to execute
other waiting operations. There are at least 2 important consequences:

1. `putAll` may fail in the middle of execution due to partition
migration. In such case it will be reexecuted entirely (only migrated
partition) on the new owner thank to invocation retry logic. Already
done items will be treated as updates of existing items which is more
costly than inserting new items.
2. CPU cache trashing if we execute putAll for many partitions
concurrently (which is expected). Their executions will be interleaved.
This might not be a big problem if the single execution slice is allowed
to take hundreds of milliseconds or more. Possibly it would be better to
finish one partition before starting next one, but still allowing
execution of other partition operations. For this we would need to have
a much more sophisticated vector operation scheduler.

I reused `metWithPreconditions` checks, but for that I had to access
some deeply buried objects. IMap StepEngine does the same.

GitOrigin-RevId: 41011b50a34fbe60f0ae21a51ab11104e1d9d003