Re-introduce RU code for 5.2 -> 5.5 upgrade [DEX-66] (#1383)
This PR reverts parts of the following PRs related to the `5.2 -> 5.3`
upgrade.

https://github.com/hazelcast/hazelcast/pull/24595
https://github.com/hazelcast/hazelcast-enterprise/pull/6032

This, in combination with the already present code for `5.3 -> 5.5`
upgrade allows the direct upgrade from `5.2 -> 5.5`.

There was an issue with TaskDefinition, essentially caused by the
version check present in the serialization code, but `Versioned`
interface missing on the class. See comments in the PR.

GitOrigin-RevId: b22caebb21e9fc57425835286de200ed3566e25f