`JobConfig#add` not using derived `id` for some resources [HZ-4712] (#1717)
In https://github.com/hazelcast/hazelcast-mono/pull/217, `JobConfig#add`
was changed to allow the `id` parameter to be `nullable`, and derived as
required.

When storing the resources, the now-`nullable` `ID` was being used
instead of the derived value.

This means that for `JobConfig`s adding resources from:
- JAR
- JARs in ZIP
- Classpath Resource by URL
- File by URL

They are not being properly stored in `JobConfig#resourceConfigs`.

Fixes: [HZ-4712](https://hazelcast.atlassian.net/browse/HZ-4712),
https://github.com/hazelcast/hazelcast/issues/26320

[HZ-4712]:
https://hazelcast.atlassian.net/browse/HZ-4712?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Aleksandar Peychev <154250829+AleksPeychev@users.noreply.github.com>
GitOrigin-RevId: efc0926334bc5f0470e1f2a739f9d3a3ef3a40a2