Fix testCallbackAfterServerShutdown flaky test  (#1772)
The logs indicate that the client is closed before
makeSureDisconnectedFromServer runs, and therefore
"getHazelcastClientInstanceImpl(client)" in
makeSureDisconnectedFromServer returns null and leads to a NPE. We need
more timeout for this test.

I changed the timeout from 5 seconds to 15 seconds, the test requires
the client to be shutdown due to cluster timeout. So we can't simply
remove the timeout. This PR will be ported to 5.2, 5.4 and master as
well.

The logs:
[report.log](https://github.com/hazelcast/hazelcast-mono/files/15197477/report.log)

Fixes https://github.com/hazelcast/hazelcast-mono/issues/1559

GitOrigin-RevId: 6bb751e9d3ca7fbb5e722f6ca9c3b7e8a5d79f0e