Make XSD links consistent with declared `XLMNS` (#1652)
The XLMNS of the Hazelcast schemas is in the format
`http://www.hazelcast.com/schema`, but _some_ links to the schemas
reference `https://hazelcast.com` (i.e. `https`, no `www`).

This was first introduced in 97aa6833c321436ed65d677dc5f91bcf8ee58ed7
and has been spread outwards since then.

Logically this is the same but it makes sense for the documentation to
actually reference the same target as the schema itself.

GitOrigin-RevId: 7e494fcd24e24c77c107892a2d4f586ec7ba3ae0