Stop overwriting `ClusterProperty.ENTERPRISE_LICENSE_KEY` when executing parallel test children (#1864)
`TlsFunctionalTest` is [intermittently
failing](https://github.com/hazelcast/hazelcast-mono/issues/1455) with
an `InvalidLicenseException`.

Further investigation showed that although the test explicitly sets a
`V6_UNLIMITED_PLATFORM_LICENSE`:

https://github.com/hazelcast/hazelcast-mono/blob/0a908b2dd9c6b389c5afbc01ecdbe2146bea3a22/hazelcast-enterprise/src/test/java/com/hazelcast/client/nio/ssl/TlsFunctionalTest.java#L126

_Some_ of the time the instances were using a
`UNLIMITED_PLATFORM_LICENSE`.

When a child parallel test is run from
`EnterpriseParallelJUnitClassRunner#runChild`, it sets the license to a
default value:

https://github.com/hazelcast/hazelcast-mono/blob/0a908b2dd9c6b389c5afbc01ecdbe2146bea3a22/hazelcast-enterprise-test-support/src/main/java/com/hazelcast/enterprise/LicenseUtil.java#L12

But in this case, it's *overwriting* the `V6_UNLIMITED_PLATFORM_LICENSE`
set earlier.

I would _assume_ that the setup in `@Before` is executed _after_ the
runner (and thus default license property) is instantiated, so this
issue does not occur. But in a parallel execution, the instantiation of
the _next_ runner could affect the license available for other,
uncompleted tests.

In any case, it seems sensible to only set the `ENTERPRISE_LICENSE_KEY`
`ClusterProperty` if it's _not already_ set to avoid overwriting.

This issue has been present for a long time, but is _likely_ more
visible now that the TPC functionality requires a special license not
under `UNLIMITED_PLATFORM_LICENSE`.

Fixes: https://github.com/hazelcast/hazelcast-mono/issues/1455
GitOrigin-RevId: 430f247ddc60b3efa829c15b189f015db2c6598d