Require non-null VectorDocument value and vectors (#1890)
Add missing validations in Java API to be in line with assumptions and
client protocol.

Implementation detail: null is returned by `RecordStore` if the entry
does not exists. Support of null values in `RecordStore` would requires
additional logic. This is additional reason, apart from general API
design, why nulls are not allowed.

GitOrigin-RevId: 4e2c5389247b64991932023ffaa91e6ec0f5bf24