Implement ClientClusterService#getClusterVersion() [CORE-129] (#871)
Clients may need to know the cluster version if an invocation requires version-specific preparation. This is exemplified by ClientJobProxy#invokeGetJobFuture()[1], where the job completion future is obtained in a synchronous manner by the aid of events to signal the existence of the job. The proxy needs to know whether the cluster supports synchronous job future retrieval to register the event listener and wait on response.

This changeset does not implement ClientClusterProxy#getClusterVersion(), which is a public API.

Protocol changes: https://github.com/hazelcast/hazelcast-client-protocol/pull/511
Relates to https://github.com/hazelcast/hazelcast/issues/19294
Closes https://hazelcast.atlassian.net/browse/CORE-129

[1]:
https://github.com/hazelcast/hazelcast-mono/blob/5cf0e23e730527a7922028f03bfa67d32f872dcd/hazelcast/hazelcast/src/main/java/com/hazelcast/jet/impl/ClientJobProxy.java#L164-L197

GitOrigin-RevId: 1ca4d14c8674585850e9062e3e850ee9f44a59a7