Add 2-stage offloaded vector search [HZAI-36] [HZAI-51] [HZAI-62] (#1415)
Adds 2-stage vector search and vector search operations offloading from
partition thread.

2-stage search sends 1 operation to each member to obtain
partially-aggregated results from the partitions owned by the member and
then aggregates partial results to get final result. In case of error
(in particular partition migrations) partition search is retried
individually on each failed partition. 2-stage search is a default
algorithm for clusters with more than 1 data member.

`memberLimit` hint is provided to control how many results are fetched
from each member. By default `SearchOptions.limit` results are fetched.
Rationale and tradeoffs for this hint are similar to `partitionLimit`
hint. Additional twist for `memberLimit` is that, especially during
migrations or member leaving/joining, partition distribution may be not
uniform which may subtly affect the quality of the results. Partitions
for `partitionLimit` on the other hand are expected to be more uniform
unless explicit partitioning strategy is used.

1-stage search is updated to support offloading and is kept as a default
algorithm used in case of 1-member cluster. It can also be enabled using
`singleStage` hint on bigger cluster. This is provided in order to
compare how it performs. In particular, it might give better latency
than 2-stage search but at the price of increasing inter-member
communication. If that turns out to be true, 1-stage search might be
useful in latency-sensitive cases.

Main execution of vector search is offloaded to `hz:query`
[executor](https://hazelcast.atlassian.net/wiki/spaces/EN/pages/4930830359/Vector+index+storage+and+search+implementation#Search-operation-executor)
(the same that is used by IMap if
`hazelcast.query.predicate.parallel.evaluation` is enabled). Default
settings of that executor however may be not optimal for vector search
(eg. default 16 threads independent on number of CPU cores). It is
recommended to adjust these settings. The advantage of reusing the same
executor is that it is already exposed via metrics.

Offloaded search is ready to be made migration-tolerant, but that will
be implemented separately.

Collections underpinning vector index supporting structures were changed
to thread-safe, where needed. JCTools hash maps were used for the main
structures. Where needed explicit retry during search was implemented if
it is not possible to generate full result without locking. After
exceeding number of retries, we ignore errors and may return incomplete
results. This however is extremely unlikely. Other approaches (see
comments in code and in the
[TDD](https://hazelcast.atlassian.net/wiki/spaces/EN/pages/4930830359/Vector+index+storage+and+search+implementation#Concurrent-search-and-updates))
may be considered later. SearchOperation is generally quite fault
tolerant already (as a regular HZ invocation), which might mask or fix
some other `ConcurrentModificationException`s.

Unrelatedly `setAsync` in member API implementation was fixed to use
appropriate operation.

---------

Co-authored-by: Vassilis Bekiaris <vbekiaris@gmail.com>
GitOrigin-RevId: 61a34c0b5b2db066173933eaefa37af84a64cd84