Add more detailed logging to cluster wide shutdown (k8s) [HZG-4] (#1950)
We need to be aware of a bit more information when the cluster wide
shutdown is performed, most notably how long it takes for the cluster
wide partition replica backup sync takes to complete with
success/failure. This can help determine a number of timeouts for the
system: k8s and member graceful shutdown periods, and cluster wide
partition replica sync timeout.
GitOrigin-RevId: 223457e6c455ecfe5e78ef15212804f558e7f46b