[HZAI-37] Adds license check for vector collection (#1737)
Adds a license check in `VectorCollectionServiceDescriptorProvider` so
that if the license is at least version 7, a usable `VectorCollection`
service is created. Otherwise, the constructed service throws
`InvalidLicenseException`. The license check is required to unblock
implementing [HZAI-63] packaging the vector search module in
`hazelcast-enterprise` artifact.
When the license extractor utility adds support for a dedicated feature
("ADVANCED AI") for vector search, the license check will be updated to
check for the specific feature.
In order to avoid circular dependency on `hazelcast-enterprise`
artifact, the license string from `Config` object is used. The license
key is set in `Config` object by `EnterpriseNodeExtension#beforeStart`,
if no OEM license is in place. In order to check for OEM license, if no
license string is found in `Config` object, we attempt to obtain the
license from `EnterpriseNodeExtension#getLicense`.

TODO in further PRs:
- Check for vector search feature instead of just license version is at
least 7 when the feature is added in `license-extractor`
- Introduce test that `VectorCollectionServiceDescriptorProvider` checks
correctly for OEM licenses when `hazelcast-enterprise-vector` module is
included in `hazelcast-enterprise`. These tests cannot be added now as
`hazelcast-enterprise-vector` does not have access to
`EnterpriseNodeExtension`.

[HZAI-63]:
https://hazelcast.atlassian.net/browse/HZAI-63?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: d982506552017bc70154cd62410cdd9889398baf