Implement vector collection security [HZAI-34] (#1960)
Vector collection permissions should be similar to other data structure
permissions like IMap or CPMap. Permissions are implemented for Java API
and [Jet](https://github.com/hazelcast/hazelcast-mono/pull/1384).

For [vector collection
optimization](https://github.com/hazelcast/hazelcast-mono/pull/1640) a
dedicated action constant is created. Index optimization is a
compute-heavy operation and during that operation collection
modifictions are not allowed. Because of that it should be more
restricted than a regular "put" action.

There a few differences in the implementation with other data
structures:

1. `VectorCollection` is not created using method in
`HazelcastInstance`. `SecureCallable` support was implemented on
`getDistributedObject` method level.
2. `VectorCollectionConfiguration` must be explicitly defined, there is
no default configuration. This affects some tests.

As a side effect `ExecutorSecurityTestSupport` was updated to be more
robust in case of lack of permissions - checks all actions instead of
failing on first one.

GitOrigin-RevId: cfe5db342553a5f53777627960aa723db29c5169