Fail fast on no memberGroups returned with SUBSET enabled [CORE-127] (#1992)
When a cluster is licensed for OS, `SUBSET` routing with
`PARTITION_GROUPS` is an unsupported mode. The connected member will
return no partition groups in this case.

It was decided that the client should fail fast and shutdown with an
`InvalidConfigurationException` , warning the user that the client has
been misconfigured.

This PR adds tests for both normal and async start of the Client.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 3bb47972fdb06a9f6f84174362c29307ce7c82ca