Resolve Kotlin dependency convergence (#2061)
[It was
identified](https://hazelcast.slack.com/archives/C01JU7ZJYGP/p1716485422610709)
that when the `docker` profile is enabled, a dependency convergence was
identified causing a build failure.

This PR:
- Addresses the convergence by:
   - `exclude`-ing one the converging transitive dependencies
   - normalises the Kotlin versions using `dependencyManagement` + `bom`
- Reverts a previous fix, [pushed in
error](https://github.com/hazelcast/hazelcast-mono/pull/2012#issuecomment-2129394547)

This is an _alternative implementation_ to
https://github.com/hazelcast/hazelcast-mono/pull/2012 which does not
actually upgrade Kotlin. From a history POV, it may be neater to resolve
the issue and upgrade the version separately.

GitOrigin-RevId: 55c2fd83f9e910217c2d6809135913196c9aa0be