Minimum cluster version checks for subset routing [CORE-127] (#2056)
The new `SubsetRouting` feature will require a minimum cluster version
of 5_5. Mixed clusters can not be supported if containing versions below
5_5 due to the required information sent in the K,V bag being absent. As
such we can not guarantee that we will be able to operate the feature in
cases such as failover to older members.

This PR restricts the minimum cluster version to 5_5, failing the client
fast with a `ClientNotAllowedInClusterException` in the case of
validation failures. The error message is updated to inform the users
that a mixed cluster with versions less than 5_5 is not permitted.

Automated compatibility testing has been added to verify functionality.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 2256ddc3baab7fe2af5aee377db083e26f9075da