Fix attribution-maven-plugin behavior in multimodule builds [DI-128] (#2072)
The `attribution-maven-plugin` in version `1.2.1` caused unresolved
artifacts files to be present in the `MavenProject` artifacts. It
sometimes led to `NullPointerException` in other plugins when a parallel
build was used.

The behavior was fixed on the plugin side:
* https://github.com/hazelcast/attribution-maven-plugin/pull/75
* https://github.com/hazelcast/attribution-maven-plugin/pull/76

This commit also takes action to speed up standard builds by
skipping the full attribution file generation. The proper attribution
file content is only generated during deployments and releases.

---------

Co-authored-by: Jack Green <jack.green@hazelcast.com>
GitOrigin-RevId: 650bd7eab7ea424e2e0297843573dcd9939ccc51