Moved Subset routing code/tests to EE [CORE-127] (#2097)
This PR moves the majority of `subset routing` specific logic and tests
to the EE code base where possible.

A new Enterprise based implementation of `ClientClusterService` extends
the base OS implementation and contains the exclusive logic to updating
and returning subset information. The class is loaded via a service
loader of EE jars are present.

`SubsetClientRoutingPartitionGroupTest` has been moved to EE code base.

The `ClientClusterService` testing has been extracted to an abstract
base class, providing a implementation for both OS and EE, with some new
tests added to the EE specific version.

In addition, new testing has been added for `SubsetMemberImpl` under
`SubsetMembersImplTest` as there was missing coverage of this class.

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 79acb587b502c74e3ecb7eb5d1c7e1e161601580