Remove executor from ClientConnectionProcessListenerRunner [HZG-25] (#2068)
The test FailoverTest.testFailover_clientDoesNotTryMemberListAfterSwitch
is flaky and frequently failing (takes 10-20 invocation to reproduce)
for a long time. The test fails because our connection listener runner
runs listeners async via an executor and we are expecting in this test
to run them synchronously.

When a address1 connection attempt is delayed a bit, and
changeClusterLatch's count becomes zero, and afterwards the listener
callback runs; it'll increase the address1 counter by 1 and the test
will fail.

Some failure investigations and logs:

This is the diff I applied before deleting the test and run the test
with it:
[a.txt](https://github.com/hazelcast/hazelcast-mono/files/15437102/a.txt)

Logs of the test when it passes:
[success2.txt](https://github.com/hazelcast/hazelcast-mono/files/15437096/success2.txt)
Logs of the test when it fails:
[failed2.txt](https://github.com/hazelcast/hazelcast-mono/files/15437098/failed2.txt)

As can be seen in the failure logs, the counter is incremented after the
client resets its local state (after the cluster change latch is counted
down)

Should be backported to 5.4 because the test exists there.

Closes #1552

GitOrigin-RevId: c29822fd6546e78e4d2f2f29b25c8afc4ba3c8d7