Fix RoutingMode reporting for <5.5 clients [CORE-127] (#2167)
We are not currently making any considerations for < 5.5 clients that do
not send `RoutingMode` information - it is currently being interpreted
as `UNISOCKET` due to using the default byte value `0`.

This commit introduces a new enumeration value, `UNKNOWN` to represent
this state of outdated client routing modes.

Fixes subtask of https://hazelcast.atlassian.net/browse/CORE-127

GitOrigin-RevId: a5ca49660d9a1512724d5dbbd3827eb05ff6ca9e