VectorIndexHolder  implemented + clear method (#2145)
client protocol:
https://github.com/hazelcast/hazelcast-client-protocol/pull/524

This PR is an attempt to refactor the vector collection storage +
implementation of clear() method.

**Clear() method (https://hazelcast.atlassian.net/browse/HZAI-32)**

A new method, clear(), has been added to the vector collection storage.
This method recreates the index(es), with the expectation that the old
indexes will be deleted by the gc. The reason for this approach is to
ensure consistent search results during the clear operation.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: 01f5a0f05d3e99a2a86940c80eb961839d1d81ff