Resolve `ClusterProperty` erroneous `...needs unchecked conversion to conform to @Nonnull String` warning (#2165)
In `NLCModeLicenseTest`, Sonar (at least via Eclipse) flags that
`HazelcastProperty.getName()` _could_ be `Null`, but
`Config.setProperty(String, String)` requires non-`Null`.

It's not actually possible for this to happen thanks to
`checkPreconditions.checkHasText`, but Sonar can't reason that.

Refactored the code (by inlining + using return value over throwing an
exception) so that it can, and applied the same change in other places
as well.

GitOrigin-RevId: e2192d7fed90de6afd37c97688437d3825ae5f5d