Fix CME causing Client to shutdown [HZ-4782] (#2234)
This PR addresses a CME which can occur on removal of the final active
connection from a Client instance with `Subset` Routing enabled.

The issue presents very infrequently and it is due to the multi-thread
read/write actions on `SubsetMembersImpl#SubsetMembersView`.

Explanation -

`triggerClusterReconnection();` submits a task to an internal executor
which calls `toMemberListFromMemberGroup`. `t1`

_Whilst iterating the `Set<UUID> uuids `..._

```
Set<UUID> uuids = subsetMembersView.members();
        List<Member> members = new ArrayList<>(uuids.size());
        for (UUID uuid : uuids) {
            Member member = snapshot.members().get(uuid);
            if (member != null) {
                members.add(member);
            }
        }
```

... the underlying set members() is modified by `t2`

`SubsetMembersImpl#onConnectionRemoved` is run on `t2` and concurrently
modifies the `SubsetMembersView` ....

```
 SubsetMembersView subsetMembersView = subsetMembersViewByClusterUuid.computeIfPresent(clusterUuid,
                (uuid, current) -> {
                    Set<UUID> members = current.members();
                    members.remove(remoteUuid);
                    return members.isEmpty() ? null : current;
                });

```

This causes the CME and as a result the client is shutdown immediately
due to the `Throwable`. The original error message was `'null'`,
therefore the code has been updated to log the actual throwable for
future help debugging issues.

The chosen solution here is to refresh the `SubsetMembersView` in
sequence with removing it from the `activeConnections` collections,
before submitting any asynchronous tasks.

fixes : https://github.com/hazelcast/hazelcast-mono/issues/2227

Checklist:
- [X] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [X] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [X] Request reviewers if possible
GitOrigin-RevId: 2d6cd5a102f0d523e6541b7aef9eb2007877d411