Refactorings for high numbers of maps  [ESC-9] (#2082)
These issues appear when we have high numbers of maps in a partition.

**Issue**
Heavy CPU usage during migration when merkle tree is enabled.

**Fix**
Refactored merkle tree enabled data-structure names finder method to
iterate over namespaces instead of all record-stores created. Because it
was iterating over all record-stores to find a single matching
namespace. And during migration when there is 5000 maps and if you are
searching for a single name space, there will be 5000*5000 looping for a
single partition migration. This was causing heavy CPU usage for no
reason.

**Refactoring-1**
`getFragmentVersions` method is refactored, because it was causing so
much allocations because of capturing lambda
https://github.com/hazelcast/hazelcast-mono/pull/2082/files#diff-0a456a66a9bb2933c4781aaffc35e61e5d780ebd5d5e30a62f39ce8caf57d302R85

**Refactoring-2**
Used List type instead of Set:
https://github.com/hazelcast/hazelcast-mono/pull/2082/files#diff-d312a0cd512b36034a77df770f81d812fc0719dd2cbde5af305c2200bc7b05f4R58

**Refactoring-3**
Refactored `shouldEnableMerkleTree` method to use existing mapContainer:
https://github.com/hazelcast/hazelcast-mono/pull/2082/files#diff-0c494dabadecc16a8cb4c7d2586b6916f6bff4a2d6008c3cfe65d595422fa687R393

GitOrigin-RevId: 199cdfcd0065f569356c388773f71a0f7b8a3f05